// apps/incare/src/modules/shared/enums/system.params.enum.ts
export enum SystemParams {
  Languages = 'config_supported_languages',
  Categories = 'config_news_categories',
  ValidPhoneCountryCodes = 'config_valid_phone_country_codes',
  AppAiRequestOption = 'config_app_ai_request_option',
  AppAiModelUrl = 'config_app_ai_model_url',
  ProductInfo = 'config_product_info',
}